refactor: api-response.ts & api-error.ts로 파일 분리 및 네이밍변경
